requirements.txt
pydub
pyaudio
asyncio
reactivex
agno
torch
faster-whisper
google-genai
librosa
kokoro-onnx
sounddevice
